Addressed recommended changes
